Create assertWithinRange
Write a function called assertWithinRange. It will be a function that takes four parameters: low a number value, which will be the lower bound on our range; high, also a number value, which will be upper bound on our range; actual, which will be a number value, and should ideally be the result of calling a function that you are testing. (presumably the function being tested should return a number value). Finally, testName will be a string, and should describe what a call to assertObjectsEqual is verifying about the function being tested.