Create Applying assertEqual 1
Write an "assertEqual" function once more, from memory. Then use your assertion function in a series of test cases to thoroughly test the code. Use categorical reasoning to consider all the useful cases. The only automated test will be one to ensure that you have created a function that is named assertEqual. It is up to you to construct the rest of the "test suite".