Create Applying assertEqual 2
The code under test is a function called every. every is an example of a "higher order" function, meaning that it takes as a parameter, another function (higher order functions can also return functions as values). This function takes two parameters; the first will be an array of scalar values, and the second parameter will be callbackFunction. Thus, when you are creating a test suite to verify that every works, you are going to need a function that you will use as an argument when you call every. Think about it like this, every is going to iterate over an array of scalar values, and apply a function to every value. This function MUST return a boolean value, mainly because we want every to tell us whether or not, EVERY value in the array returns true when the callbackFunction is called upon it.